server/requirements.txt
fastapi
uvicorn[standard]
livekit
python-dotenv
pydantic
PyJWT
livekit-api
asyncpg
databases[postgresql]
sqlalchemy[asyncio]
alembic
python-jose[cryptography]
passlib[bcrypt]
psycopg2-binary
bcrypt
gunicorn
torch
transformers
librosa
soundfile
openai
python-multipart